Fixed compiler warnings.
git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4585 b9700a98-9937-0410-b777-f318d6eb5bff

+++ Tracking/include/QwRayTracer.h
@@ -96,7 +96,7 @@ class QwRayTracer: public VQwBridgingMethod {
     static QwMagneticField *fBfield;
 
     /// Runge-Kutta method order
-    double fIntegrationOrder;
+    int fIntegrationOrder;
     /// Runge-Kutta step size
     double fIntegrationStep;
